Tests: Cards

+++ src/entities/Card.java
@@ -58,7 +58,7 @@ public class Card {
         return cardNumber;
     }
 
-    public void setCardNumber(int cardNumber) {
+    public void setCardNumber(long cardNumber) {
         this.cardNumber = cardNumber;
     }
